Move guidance-links to common so we can reuse it
